+++ src/SmartHopper.Core.Grasshopper/Utils/GHCanvasUtils.cs
@@ -12,6 +12,7 @@ using Grasshopper;
 using Grasshopper.Kernel;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Drawing;
 using System.Linq;
 
@@ -35,7 +36,7 @@ namespace SmartHopper.Core.Grasshopper.Utils
         }
 
         // Add an object to the canvas
-        public static void AddObjectToCanvas(IGH_DocumentObject obj, PointF position = default, bool live = false)
+        public static void AddObjectToCanvas(IGH_DocumentObject obj, PointF position = default, bool redraw = true)
         {
             GH_Document doc = GetCurrentCanvas();
 
@@ -43,8 +44,9 @@ namespace SmartHopper.Core.Grasshopper.Utils
 
             doc.AddObject(obj, false);
 
-            if (live)
+            if (redraw)
             {
+                obj.Attributes.ExpireLayout();
                 Instances.RedrawCanvas();
             }
         }
@@ -70,6 +72,31 @@ namespace SmartHopper.Core.Grasshopper.Utils
                 //Debug.WriteLine("The object is neither an IGH_Component nor an IGH_Param.");
                 return obj;
             }
+        }
+
+        /// <summary>
+        /// Moves an existing instance by setting its Pivot position by GUID.
+        /// </summary>
+        /// <param name="guid">The GUID of the instance to move.</param>
+        /// <param name="position">The new pivot position, absolute or relative.</param>
+        /// <param name="relative">True to interpret position as a relative offset; false for absolute.</param>
+        /// <param name="redraw">True to redraw canvas after moving.</param>
+        /// <returns>True if the instance was found and moved; otherwise false.</returns>
+        public static bool MoveInstance(Guid guid, PointF position, bool relative = false, bool redraw = true)
+        {
+            var obj = FindInstance(guid);
+            if (obj == null) return false;
+            var current = obj.Attributes.Pivot;
+            var target = relative
+                ? new PointF(current.X + position.X, current.Y + position.Y)
+                : position;
+            obj.Attributes.Pivot = target;
+            if (redraw)
+            {
+                obj.Attributes.ExpireLayout();
+                Instances.RedrawCanvas();
+            }
+            return true;
         }
 
         // Identify occupied areas